requirements.txt
configparser==3.5.0
Django==1.10.1
djangorestframework==3.4.7
enum34==1.1.6
flake8==3.0.4
mccabe==0.5.2
pycodestyle==2.0.0
pyflakes==1.2.3
requests==2.11.1
-e git://github.com/vyscond/yql.git@5d13c392ef514b6ec777d4cba8472e697c85a16a#egg=yql
